add skeet-framework/ai
