+++ functions/skeet/src/routings/http/index.ts
@@ -1,2 +1,2 @@
 export * from './root'
-export * from './createStreamChatMessage'
+export * from './createVertexChatMessage'
